Break apart the Kubernetes Client
Make a seperate directory for clients so they are
not apart of the APB code.  This will make the clients
easy to initialize and create new ones.
